Fixed bug with 0s in blank calculation
Made sure everything was read in as flaot64, and then removed lines where all values are 0. This keeps rows that have just a few 0 values, and will take the mean of this rows as predicted. This solves the issues where the script was thrown off if there was a 0 in the blank measurement.